chore(diary-entity): MemberEntity 수정

+++ src/main/java/com/moneyyy/expressionDiary/entity/MemberEntity.java
@@ -14,11 +14,12 @@ public class MemberEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    // erd 수정하기(필수), Column name "id"로 고치기(보류)
     @Column(name = "member_id")
     private Long memberId;
 
     // 로그인 아이디
-    private int id;
+    private Long id;
 
     // 로그인 비밀번호
     private String password;
